refactor(alembic): consolidate all alumni chapter db model edits into a single alembic migration file
