--- backend/alembic/versions/032154e1b55a_added_alumnichapter_and_updated_.py
@@ -1,57 +1,0 @@
-"""added AlumniChapter and updated relationships with University
-
-Revision ID: 032154e1b55a
-Revises: b332f2a2aaef
-Create Date: 2026-07-07 17:36:04.078217
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '032154e1b55a'
-down_revision: Union[str, None] = 'b332f2a2aaef'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('universities',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('name', sa.String(length=255), nullable=False),
-    sa.Column('abbreviation', sa.String(length=32), nullable=True),
-    sa.Column('location', sa.String(length=255), nullable=True),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('name')
-    )
-    op.create_table('alumni_chapters',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('name', sa.String(length=255), nullable=False),
-    sa.Column('university_id', sa.Integer(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), nullable=True),
-    sa.ForeignKeyConstraint(['university_id'], ['universities.id'], ),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('university_id')
-    )
-    op.add_column('tournaments', sa.Column('university_id', sa.Integer(), nullable=True))
-    op.create_foreign_key(None, 'tournaments', 'universities', ['university_id'], ['id'])
-    op.add_column('users', sa.Column('university_id', sa.Integer(), nullable=True))
-    op.create_foreign_key(None, 'users', 'universities', ['university_id'], ['id'])
-    op.drop_column('users', 'university')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('users', sa.Column('university', sa.VARCHAR(length=255), autoincrement=False, nullable=True))
-    op.drop_constraint(None, 'users', type_='foreignkey')
-    op.drop_column('users', 'university_id')
-    op.drop_constraint(None, 'tournaments', type_='foreignkey')
-    op.drop_column('tournaments', 'university_id')
-    op.drop_table('alumni_chapters')
-    op.drop_table('universities')
-    # ### end Alembic commands ###
