--- backend/alembic/versions/05619ff431f0_add_chapter_membership_table_and_update_.py
@@ -1,40 +1,0 @@
-"""add chapter membership table and update user and alumni chapter
-
-Revision ID: 05619ff431f0
-Revises: 032154e1b55a
-Create Date: 2026-07-07 18:11:18.789042
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '05619ff431f0'
-down_revision: Union[str, None] = '032154e1b55a'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('chapter_memberships',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('chapter_id', sa.Integer(), nullable=False),
-    sa.Column('user_id', sa.Integer(), nullable=False),
-    sa.Column('role', sa.String(length=32), nullable=False),
-    sa.Column('joined_at', sa.DateTime(timezone=True), nullable=True),
-    sa.ForeignKeyConstraint(['chapter_id'], ['alumni_chapters.id'], ),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('user_id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('chapter_memberships')
-    # ### end Alembic commands ###
