--- backend/alembic/versions/389c01ee9678_add_uselist_and_cascade_on_delete_for_.py
@@ -1,36 +1,0 @@
-"""add uselist and cascade on delete for chapter alumni to chapter membership and chapter join code
-
-Revision ID: 389c01ee9678
-Revises: 9e59394c37a8
-Create Date: 2026-07-07 23:04:26.003398
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '389c01ee9678'
-down_revision: Union[str, None] = '9e59394c37a8'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint('chapter_join_codes_chapter_id_fkey', 'chapter_join_codes', type_='foreignkey')
-    op.create_foreign_key(None, 'chapter_join_codes', 'alumni_chapters', ['chapter_id'], ['id'], ondelete='CASCADE')
-    op.drop_constraint('chapter_memberships_chapter_id_fkey', 'chapter_memberships', type_='foreignkey')
-    op.create_foreign_key(None, 'chapter_memberships', 'alumni_chapters', ['chapter_id'], ['id'], ondelete='CASCADE')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'chapter_memberships', type_='foreignkey')
-    op.create_foreign_key('chapter_memberships_chapter_id_fkey', 'chapter_memberships', 'alumni_chapters', ['chapter_id'], ['id'])
-    op.drop_constraint(None, 'chapter_join_codes', type_='foreignkey')
-    op.create_foreign_key('chapter_join_codes_chapter_id_fkey', 'chapter_join_codes', 'alumni_chapters', ['chapter_id'], ['id'])
-    # ### end Alembic commands ###
