--- backend/alembic/versions/9e59394c37a8_add_tournament_chapter_model_and_.py
@@ -1,36 +1,0 @@
-"""add tournament chapter model and relationships for Tournament and AlumniChapter
-
-Revision ID: 9e59394c37a8
-Revises: d0204558e286
-Create Date: 2026-07-07 22:38:59.263613
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '9e59394c37a8'
-down_revision: Union[str, None] = 'd0204558e286'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('tournament_chapters',
-    sa.Column('tournament_id', sa.Integer(), nullable=False),
-    sa.Column('chapter_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['chapter_id'], ['alumni_chapters.id'], ),
-    sa.ForeignKeyConstraint(['tournament_id'], ['tournaments.id'], ),
-    sa.PrimaryKeyConstraint('tournament_id', 'chapter_id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('tournament_chapters')
-    # ### end Alembic commands ###
