--- backend/alembic/versions/b332f2a2aaef_add_university_table_and_update_user_.py
@@ -1,47 +1,0 @@
-"""add university table and update user and tournament relationships
-
-Revision ID: b332f2a2aaef
-Revises: f250cb93a643
-Create Date: 2026-07-06 00:13:05.135282
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'b332f2a2aaef'
-down_revision: Union[str, None] = 'f250cb93a643'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('universities',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('name', sa.String(length=255), nullable=False),
-    sa.Column('abbreviation', sa.String(length=32), nullable=True),
-    sa.Column('location', sa.String(length=255), nullable=True),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('name')
-    )
-    op.add_column('tournaments', sa.Column('university_id', sa.Integer(), nullable=True))
-    op.create_foreign_key(None, 'tournaments', 'universities', ['university_id'], ['id'])
-    op.add_column('users', sa.Column('university_id', sa.Integer(), nullable=True))
-    op.create_foreign_key(None, 'users', 'universities', ['university_id'], ['id'])
-    op.drop_column('users', 'university')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('users', sa.Column('university', sa.VARCHAR(length=255), autoincrement=False, nullable=True))
-    op.drop_constraint(None, 'users', type_='foreignkey')
-    op.drop_column('users', 'university_id')
-    op.drop_constraint(None, 'tournaments', type_='foreignkey')
-    op.drop_column('tournaments', 'university_id')
-    op.drop_table('universities')
-    # ### end Alembic commands ###
