--- backend/alembic/versions/d0204558e286_add_chapter_join_codes_and_update_.py
@@ -1,43 +1,0 @@
-"""add chapter join codes and update relationships
-
-Revision ID: d0204558e286
-Revises: 05619ff431f0
-Create Date: 2026-07-07 22:07:36.313956
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'd0204558e286'
-down_revision: Union[str, None] = '05619ff431f0'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('chapter_join_codes',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('chapter_id', sa.Integer(), nullable=False),
-    sa.Column('created_by', sa.Integer(), nullable=False),
-    sa.Column('code', sa.String(length=8), nullable=False),
-    sa.Column('label', sa.String(length=255), nullable=True),
-    sa.Column('expires_at', sa.DateTime(timezone=True), nullable=True),
-    sa.Column('is_active', sa.Boolean(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), nullable=True),
-    sa.ForeignKeyConstraint(['chapter_id'], ['alumni_chapters.id'], ),
-    sa.ForeignKeyConstraint(['created_by'], ['users.id'], ),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('code')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('chapter_join_codes')
-    # ### end Alembic commands ###
